# openapi/crm.openapi.yaml
openapi: 3.0.3
info:
  title: AI CRM Backend API
  version: 0.1.0
  description: >
    Skeleton OpenAPI contract for the CRM backend. This spec is intentionally
    contract-first and can be refined endpoint-by-endpoint.
servers:
  - url: https://api.yourcrm.com/v1
    description: Production
  - url: https://{apiId}.execute-api.{region}.amazonaws.com/{stage}
    description: AWS API Gateway
    variables:
      apiId:
        default: your-api-id
      region:
        default: us-east-1
      stage:
        default: dev
security:
  - bearerAuth: []
...
paths:
...
  /contacts/{id}:
    get:
      tags: [Contacts]
      operationId: getContactById
      summary: Get contact by id
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Contacts]
      operationId: updateContact
      summary: Update contact
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdateRequest"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Contacts]
      operationId: deleteContact
      summary: Delete contact
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Contact deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /companies/{id}:
    get:
      tags: [Companies]
      operationId: getCompanyById
      summary: Get company by id
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Companies]
      operationId: updateCompany
      summary: Update company
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdateRequest"
      responses:
        "200":
          description: Company updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Companies]
      operationId: deleteCompany
      summary: Delete company
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Company deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /deals:
    get:
      tags: [Deals]
      operationId: listDeals
      summary: List deals
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SearchOptionalParam"
      responses:
        "200":
          description: Deal list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDealsResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags: [Deals]
      operationId: createDeal
      summary: Create deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealCreateRequest"
      responses:
        "201":
          description: Deal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /deals/{id}:
    get:
      tags: [Deals]
      operationId: getDealById
      summary: Get deal by id
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Deal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Deals]
      operationId: updateDeal
      summary: Update deal
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealUpdateRequest"
      responses:
        "200":
          description: Deal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Deals]
      operationId: deleteDeal
      summary: Delete deal
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Deal deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /tasks/{id}:
    get:
      tags: [Tasks]
      operationId: getTaskById
      summary: Get task by id
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags: [Tasks]
      operationId: updateTask
      summary: Update task
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Tasks]
      operationId: deleteTask
      summary: Delete task
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPathParam:
      name: id
      in: path
      required: true
      schema:
        type: string
    CursorParam:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    SearchOptionalParam:
      name: q
      in: query
      required: false
      schema:
        type: string
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ValidationIssue:
      type: object
      required: [path, message]
      properties:
        path:
          type: string
        message:
          type: string
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    CustomerType:
      type: string
      enum: [B2B, B2C, PATIENT]
    DealStatus:
      type: string
      enum: [OPEN, WON, LOST]
    TaskStatus:
      type: string
      enum: [OPEN, DONE]
...
    Company:
      type: object
      required: [id, workspaceId, ownerId, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        ownerId:
          type: string
        name:
          type: string
        domain:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Contact:
      type: object
      required: [id, workspaceId, ownerId, firstName, lastName, tags, createdAt, updatedAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        ownerId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        companyId:
          type: string
          nullable: true
        customerType:
          $ref: "#/components/schemas/CustomerType"
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Deal:
      type: object
      required:
        - id
        - workspaceId
        - ownerId
        - title
        - amount
        - currency
        - stageId
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        workspaceId:
          type: string
        ownerId:
          type: string
        title:
          type: string
        amount:
          type: number
        currency:
          type: string
        stageId:
          type: string
        companyId:
          type: string
          nullable: true
        primaryContactId:
          type: string
          nullable: true
        expectedCloseDate:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/DealStatus"
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Task:
      type: object
      required:
        - id
        - workspaceId
        - createdById
        - title
        - status
        - relatedType
        - relatedId
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        workspaceId:
          type: string
        createdById:
          type: string
        assigneeId:
          type: string
          nullable: true
        title:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
        relatedType:
          type: string
          enum: [contact, company, deal, task]
        relatedId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ContactUpdateRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        jobTitle:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        phone:
          type: string
          nullable: true
        companyId:
          type: string
          nullable: true
        customerType:
          $ref: "#/components/schemas/CustomerType"
        tags:
          type: array
          items:
            type: string
...
    CompanyUpdateRequest:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
          nullable: true
        industry:
          type: string
          nullable: true
        size:
          type: string
          nullable: true
    DealCreateRequest:
      type: object
      required: [title, amount, currency, stageId]
      properties:
        title:
          type: string
        amount:
          type: number
        currency:
          type: string
        stageId:
          type: string
        companyId:
          type: string
          nullable: true
        primaryContactId:
          type: string
          nullable: true
        expectedCloseDate:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/DealStatus"
        description:
          type: string
          nullable: true
    DealUpdateRequest:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
        currency:
          type: string
        stageId:
          type: string
        companyId:
          type: string
          nullable: true
        primaryContactId:
          type: string
          nullable: true
        expectedCloseDate:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/DealStatus"
        description:
          type: string
          nullable: true
...
    TaskUpdateRequest:
      type: object
      properties:
        title:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        assigneeId:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/TaskStatus"
...
    ListDealsResponse:
      type: object
      required: [rows, total]
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/Deal"
        total:
          type: integer
        nextCursor:
          type: string
          nullable: true
...